Ran 2to3 and cleaned up: 2to3 -w -n -f all -f idioms -f set_literal -x dict -x future -x unicode -x print -x xrange gpaw

+++ gpaw/grid_descriptor.py
@@ -571,7 +571,7 @@ class GridDescriptor(Domain):
         
     def bytecount(self, dtype=float):
         """Get the number of bytes used by a grid of specified dtype."""
-        return long(np.prod(self.n_c)) * np.array(1, dtype).itemsize
+        return int(np.prod(self.n_c)) * np.array(1, dtype).itemsize
 
     def get_grid_point_coordinates(self, dtype=float, global_array=False):
         """Construct cartesian coordinates of grid points in the domain."""
